Override the Facebook platform's HTTP POST request to HTTP GET.
In March 2011, Facebook imposed a migration that changed their platform's
request method from GET to POST in an attempt to screen the signed request
contained within from being propagated through the HTTP 'referer' header.
Contrary to what the the developers at Facebook might think, though, this
is a terrible idea because HTTP GET and HTTP POST requests do very
different things. As per this commit, Fandjango-powered applications are
immune to this crime against the HTTP protocol standard.

As an added bonus, this also addresses an issue with Django's built-in
CSRF protection blocking requests from the Facebook platform.